+++ timeutil/tinydate.go
@@ -25,6 +25,10 @@ var StartDate = uint32(1322611200)
 
 // TinyDate creates tiny date struct by timestamp
 func TinyDate(t int64) Date {
+	if t < int64(StartDate) {
+		return 0
+	}
+
 	return Date(uint32(t) - StartDate)
 }