+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [13.38.8](https://kaos.sh/ek/13.38.8)
 
+- **`[fmtc]`** Fixed Off-by-One bug in 256-Color range validation
 - **`[log]`** Code refactoring
 - **`[req]`** Code refactoring